fixes osx catalina error while build

+++ src/cryptonotecore/Core.cpp
@@ -1001,7 +1001,7 @@ namespace CryptoNote
 
         uint8_t nextBlockMajorVersion = getBlockMajorVersionForHeight(topBlockIndex);
 
-        size_t blocksCount = std::min(static_cast<size_t>(topBlockIndex), CryptoNote::parameters::DIFFICULTY_BLOCKS_COUNT);
+        size_t blocksCount = std::min(static_cast<uint64_t>(topBlockIndex), CryptoNote::parameters::DIFFICULTY_BLOCKS_COUNT);
 
         auto timestamps = mainChain->getLastTimestamps(blocksCount);
         auto difficulties = mainChain->getLastCumulativeDifficulties(blocksCount);
